pyproject.toml
[build-system]
requires = ["setuptools>=67"]
build-backend = "setuptools.build_meta"

[project]
name = "autocase"
version = "0.1.0"
description = "AutoCase - LLM-friendly test case generator"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "pyyaml>=6.0",
  "openpyxl>=3.1",
  "openai>=1.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
autocase = "autocase.cli:main"
